Install babel, move the test file to the test folder
